Move meta variables from flask_monitor to setup.py
If setup.py has to import flask_monitor, then it also has to import
flask.

Which means that flask already has to be present in order to install it
(or even build the package).

The meta variables are only used by setup.py, so they should really live
in the setup.py file.